+++ dune/gdt/basefunctionset/fem.hh
@@ -28,8 +28,8 @@ namespace BaseFunctionSet {
 
 // forward, to be used in the traits and to allow for specialization
 template< class ShapeFunctionSetImp, class EntityImp,
-          class DomainFieldImp, int domainDim,
-          class RangeFieldImp, int rangeDim, int rangeDimCols = 1 >
+          class DomainFieldImp, size_t domainDim,
+          class RangeFieldImp, size_t rangeDim, size_t rangeDimCols = 1 >
 class FemWrapper
 {
   static_assert(Dune::AlwaysFalse< ShapeFunctionSetImp >::value, "Untested for these dimensions!");
@@ -40,8 +40,8 @@ namespace internal {
 
 
 template< class ShapeFunctionSetImp, class EntityImp,
-          class DomainFieldImp, int domainDim,
-          class RangeFieldImp, int rangeDim, int rangeDimCols >
+          class DomainFieldImp, size_t domainDim,
+          class RangeFieldImp, size_t rangeDim, size_t rangeDimCols >
 class FemWrapperTraits
 {
 public:
@@ -57,8 +57,8 @@ public:
 
 
 template< class ShapeFunctionSetImp, class EntityImp,
-          class DomainFieldImp, int domainDim,
-          class RangeFieldImp, int rangeDim >
+          class DomainFieldImp, size_t domainDim,
+          class RangeFieldImp, size_t rangeDim >
 class FemWrapper< ShapeFunctionSetImp, EntityImp, DomainFieldImp, domainDim, RangeFieldImp, rangeDim, 1 >
   : public BaseFunctionSetInterface< internal::FemWrapperTraits< ShapeFunctionSetImp, EntityImp,
                                                        DomainFieldImp, domainDim, RangeFieldImp, rangeDim, 1 >,
@@ -132,8 +132,8 @@ private:
 
 
 template< class ShapeFunctionSetImp, class EntityImp,
-          class DomainFieldImp, int domainDim,
-          class RangeFieldImp, int rangeDim, int rangeDimCols = 1 >
+          class DomainFieldImp, size_t domainDim,
+          class RangeFieldImp, size_t rangeDim, size_t rangeDimCols = 1 >
 class FemWrapper
 {
   static_assert(Dune::AlwaysFalse< ShapeFunctionSetImp >::value, "You are missing dune-fem!");
